Moved wpmediapicker to main UI package
