+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/wpmediapicker/WPMediaPickerFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.products.wpmediapicker
+package com.woocommerce.android.ui.wpmediapicker
 
 import android.content.DialogInterface
 import android.os.Bundle
@@ -23,7 +23,7 @@ import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.base.UIMessageResolver
 import com.woocommerce.android.ui.dialog.CustomDiscardDialog
 import com.woocommerce.android.ui.main.MainActivity.Companion.BackPressListener
-import com.woocommerce.android.ui.products.wpmediapicker.WPMediaGalleryView.OnWPMediaGalleryClickListener
+import com.woocommerce.android.ui.wpmediapicker.WPMediaGalleryView.OnWPMediaGalleryClickListener
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ShowSnackbar
 import com.woocommerce.android.viewmodel.ViewModelFactory
